check for config at /etc/waylon.yml first
Prior to this commit, we only used configuration files located at
config/waylon.yml. However, this might not be the case for production
deployments.

This commit first checks to see if a config file exists at
/etc/waylon.yml. If it does, it will be used. Otherwise, we use the
config file located at config/waylon.yml, just like we've done in the
past.