use /var/log/waylon for logging, if it exists
Prior to this commit, we only logged to the logs/ directory located in
the application root directory. This might be inconvenient for users
trying to deploy the application in production.

This commit first checks to see if /var/log/waylon exists. If it does,
logs will be placed there. Otherwise, we fallback to using the logs/
directory within the application root, and create it if it doesn't
already exist (like we've done up to this point).