unicorn config should specify working_directory
Prior to this commit, we did not use the 'working_directory' config
param in unicorn.rb. This caused problems when trying to configure and
launch Unicorn outside of the application directory, e.g. with an init
script.

This commit configures the application's working directory to be
relative to the 'unicorn.rb' config file, which we already expect to be
at $APP_ROOT/config/unicorn.rb.