(maint) Modify Waylon for development on Windows
Previously, even though ruby is cross platform, Waylon could not be used on a
Windows platform due to the use of the unicorn and memcached gems.  These gems
have native extensions which can not be compiled on Windows.  This commit
modifies Waylon so at least the majority of Waylon development can occur on
Windows:

- Modifies the Gemfile to not include those gems on Windows platforms
- Modifies Waylon to not try to require memcached based classed when on Windows
  platforms
- Adds a Procfile.windows so that foreman can be used to run Waylon on Window
- Updates documentation on how to run foreman on Windows
- Forces STDOUT synchronisation in rack.  Previously the log entries were being
  swallowed until the app quit (flushed the STDOUT buffer).  The log output from
  Waylon is small enough so this will not cause an issue