(maint) Add nicer timestamps in view
Previously the timestamps displayed on minutes and seconds, whereas hours was
also important.  Also there was no information on when the last build was
finished.  This commit modifies the timespan tet converter so that it displays
days, hours, minutes and seconds as appropriate.  Also it does not show units
which are not required e.g. `0d 0h 0m 12s` shows only as `12s` because the other
units are all zero.

This commit also adds a `since_last_build` value to the job api, which is the
amount of time since the last job was started (derived form last_build_timestamp).